+++ tests/fixtures/cases/basic-with-default-update-service/__expected/main.js
@@ -1,0 +1,136 @@
+/******/ (() => { // webpackBootstrap
+/******/ 	var __webpack_modules__ = ([
+/* 0 */,
+/* 1 */
+/***/ ((__unused_webpack_module, exports) => {
+
+var appCacheIframe;
+
+function hasSW() {
+  
+    return 'serviceWorker' in navigator && (
+      window.location.protocol === 'https:' ||
+      /^(.*\.)?localhost$/.test(window.location.hostname) || // only localhost or ends with .localhost
+      window.location.hostname.indexOf('127.') === 0
+    );
+  
+}
+
+function install(options) {
+  options || (options = {});
+
+  
+    if (hasSW()) {
+      var registration = navigator.serviceWorker
+        .register(
+          "sw.js", {
+            
+            
+          }
+        );
+
+      
+
+      return;
+    }
+  
+
+  
+    if (window.applicationCache) {
+      var directory = "appcache/";
+      var name = "manifest";
+
+      var doLoad = function() {
+        var page = directory + name + '.html';
+        var iframe = document.createElement('iframe');
+
+        
+
+        iframe.src = page;
+        iframe.style.display = 'none';
+
+        appCacheIframe = iframe;
+        document.body.appendChild(iframe);
+      };
+
+      if (document.readyState === 'complete') {
+        setTimeout(doLoad);
+      } else {
+        window.addEventListener('load', doLoad);
+      }
+
+      return;
+    }
+  
+}
+
+function applyUpdate(callback, errback) {
+  
+
+  
+}
+
+function update() {
+  
+    if (hasSW()) {
+      navigator.serviceWorker.getRegistration().then(function(registration) {
+        if (!registration) return;
+        return registration.update();
+      });
+    }
+  
+
+  
+    if (appCacheIframe) {
+      try {
+        appCacheIframe.contentWindow.applicationCache.update();
+      } catch (e) {}
+    }
+  
+}
+
+
+  setInterval(update, 3600000);
+
+
+exports.install = install;
+exports.applyUpdate = applyUpdate;
+exports.update = update;
+
+
+/***/ })
+/******/ 	]);
+/************************************************************************/
+/******/ 	// The module cache
+/******/ 	var __webpack_module_cache__ = {};
+/******/ 	
+/******/ 	// The require function
+/******/ 	function __webpack_require__(moduleId) {
+/******/ 		// Check if module is in cache
+/******/ 		var cachedModule = __webpack_module_cache__[moduleId];
+/******/ 		if (cachedModule !== undefined) {
+/******/ 			return cachedModule.exports;
+/******/ 		}
+/******/ 		// Create a new module (and put it into the cache)
+/******/ 		var module = __webpack_module_cache__[moduleId] = {
+/******/ 			// no module.id needed
+/******/ 			// no module.loaded needed
+/******/ 			exports: {}
+/******/ 		};
+/******/ 	
+/******/ 		// Execute the module function
+/******/ 		__webpack_modules__[moduleId](module, module.exports, __webpack_require__);
+/******/ 	
+/******/ 		// Return the exports of the module
+/******/ 		return module.exports;
+/******/ 	}
+/******/ 	
+/************************************************************************/
+var __webpack_exports__ = {};
+// This entry needs to be wrapped in an IIFE because it needs to be isolated against other modules in the chunk.
+(() => {
+__webpack_require__(1);
+})();
+
+/******/ })()
+;